improved filters on multiple measures
For asthma_01, respiratory_02, and hypoglycemia_01, filters were improved so the initial population does not look at all patients, but only includes the lowest age limit and greater.  This makes the peds filter a bit redundant, but in this way the final N of the population of interest will be the sum of adult and peds.  Previously, the code looked at all pateints for the total population, which means that population included patients younger than the lower age limit of peds.  Fixed.