creation of the first *_population functions
the *_population family of functions are the workhorses of the nemsqar package that will complete all the calculations throughout the package.

changing the data strategy to only use the star schema framework and create dimension tables from the dataframe, or utilize the dataframes from the different tables passed to the new table functionality for each function.

All functions in the future will support passing a single dataframe or multiple tables and will produce identical results.  The same data passed to df that matches the tables approach will produce identical results.