created safety_01 and safety_01_population
several updates to the other existing functions that have been finished with the *_population modular function.

Updates include cleaning up the progress bar for the wrapper functions.  this cleanup means that the only progress bar happening is the *_population progress bar, the wrapper no longer has its own.

additionally, the only error checks happening are those happening within the modular function, the wrapper no longer has error checks as they would just be the same and that is unhelpful / duplicative.  The wrapper function only exists with the modular function anyway, so it works well, and is more efficient while avoiding errors that can be a product of repetition.