fix to trauma_14_population for patch, formatting
trauma_03 and trauma_04 were only reviewed and formatted using the Air package, no substantive changes beyond code formatting

trauma_14_population had a fix applied to the transport_data object creation section where it lacked the dplyr::if_any wrapper to grepl() when finding transport responses.  In this way, the function only allowed using one column for the transport disposition, when more were needed given changes to NEMSIS 3.4 to 3.5 which created a new column for transport dispositions.  With the fix to be released in 1.1.1, trauma_14_population will correctly accept an arbitrary number of columns in the transport_disposition argument.

updated the cran-comments, NEWS and DESCRIPTION with corresponding version update and changes